% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_asap_simple.R
\docType{data}
\name{asap_simple}
\alias{asap_simple}
\title{Age Structured Assessment Program (ASAP) input and output data from the simple example.}
\format{
A data list with 2 lists of data:
\describe{
\item{input}{ASAP stock assessment input data list. It includes 4 lists of data: dat, comments, fleet.names, and survey.names.}
\item{output}{ASAP stock assessment output data list. It includes 58 lists of data.}
...
}
}
\source{
\url{https://nmfs-fish-tools.github.io/ASAP/}
}
\usage{
asap_simple
}
\description{
A dataset containing the ASAP stock assessment input and output for a simple example.
}
\keyword{datasets}
