% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asap_simple_input.R
\docType{data}
\name{asap_simple_input}
\alias{asap_simple_input}
\title{Age Structured Assessment Program (ASAP) input data from the simple example.}
\format{
A data list with 4 lists of data:
\describe{
\item{dat}{stock assessment inputs from ASAP input data file}
\item{comments}{comments from ASAP input data file}
\item{fleet.names}{name of each fleet}
\item{survey.names}{name of each survey}
...
}
}
\source{
\url{https://nmfs-fish-tools.github.io/ASAP/}
}
\usage{
asap_simple_input
}
\description{
A dataset containing the ASAP stock assessment inputs for a simple example.
}
\keyword{datasets}
