% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asap_simple_output.R
\docType{data}
\name{asap_simple_output}
\alias{asap_simple_output}
\title{Age Structured Assessment Program (ASAP) output data from the simple example.}
\format{
A data list with 58 lists of data:
\describe{
\item{info}{date of the outputs and program name}
...
}
}
\source{
\url{https://nmfs-fish-tools.github.io/ASAP/}
}
\usage{
asap_simple_output
}
\description{
A dataset containing the ASAP stock assessment outputs for a simple example.
}
\keyword{datasets}
