% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishsad-package.R
\docType{package}
\name{fishsad}
\alias{fishsad}
\title{fishsad: A package for fisheries stock assessment data}
\description{
The fishsad package provides input and output data from multiple stock assessment packages. The data can be used for model testing and comparison.
}
