% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ss_empiricalwaa.R
\docType{data}
\name{ss_empiricalwaa}
\alias{ss_empiricalwaa}
\title{Stock Synthesis (SS) input data from the empirical_watage_and_age_selex example.}
\format{
A data list with 1 list of data:
\describe{
\item{input}{SS stock assessment input data list. It includes 5 lists of data: ss_data, ss_control, ss_starter, ss_forecast, ss_wtatage.}
...
}
}
\source{
\url{https://github.com/nmfs-stock-synthesis/user-examples}
}
\usage{
ss_empiricalwaa
}
\description{
A dataset containing the SS stock assessment input.
}
\keyword{datasets}
