% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_empiricalwaa_input.R
\docType{data}
\name{ss_empiricalwaa_input}
\alias{ss_empiricalwaa_input}
\title{Stock Synthesis (SS) input data from the empirical_watage_and_age_selex example.}
\format{
A data list with 5 lists of data:
\describe{
\item{ss_starter}{input data for ss_starter}
\item{ss_wtatage}{input data for ss_wtatage}
\item{ss_control}{input data for ss_control}
\item{ss_data}{input data for ss_data}
\item{ss_forecast}{input data for ss_forecast}
...
}
}
\source{
\url{https://github.com/nmfs-stock-synthesis/user-examples}
}
\usage{
ss_empiricalwaa_input
}
\description{
A dataset containing the SS stock assessment inputs.
}
\keyword{datasets}
