% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ss_input.R
\name{ss_input}
\alias{ss_input}
\title{Read SS input files}
\usage{
ss_input(
  file_path,
  data_file,
  control_file,
  starter_file,
  forecast_file,
  watage_file,
  version = "3.30"
)
}
\arguments{
\item{file_path}{Path to input files}

\item{data_file}{Name of data file}

\item{control_file}{Name of control file}

\item{starter_file}{Name of starter file}

\item{forecast_file}{Name of forecast file}

\item{watage_file}{Name of weight-at-age data file}

\item{version}{SS version number}
}
\value{
Returns a list with input data.
}
\description{
Read in data, control, starter, forecast, weight-at-age data files into a data list in R.
}
